pyproject.toml
# Copyright 2022-2025 Rigetti & Co, LLC
#
# This Computer Software is developed under Agreement HR00112230006 between Rigetti & Co, LLC and
# the Defense Advanced Research Projects Agency (DARPA). Use, duplication, or disclosure is subject
# to the restrictions as stated in Agreement HR00112230006 between the Government and the Performer.
# This Computer Software is provided to the U.S. Government with Unlimited Rights; refer to LICENSE
# file for Data Rights Statements. Any opinions, findings, conclusions or recommendations expressed
# in this material are those of the author(s) and do not necessarily reflect the views of the DARPA.
#
# Use of this work other than as specifically authorized by the U.S. Government is licensed under
# the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance
# with the License. You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software distributed under the License
# is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
# or implied. See the License for the specific language governing permissions and limitations under
# the License.

[tool.black]
line-length = 120
exclude = "external/"

[tool.coverage.report]
exclude_also = [
    # Exclude non-implemented methods of abstract base classes that are not meant to be reachable.
    "raise NotImplementedError",
    # Exclude run-time errors, which are used defensively for conditions that should not occur.
    "raise RuntimeError",
    # Exclude exceptions
    "except AssertionError",
    "except Exception",
]

[tool.poetry]
name = "rigetti_resource_estimation"
version = "0.26.32"
description = "A resource estimator for future fault-tolerant superconducting quantum hardware based on graph state compilation to surface code widgets."
license = "As per Agreement HR00112230006 for the US Government; Apache 2.0 otherwise."
authors = ["Mark Hodson <mhodson@rigetti.com>", "Josh Mutus <jmutus@rigetti.com>", "Nariman Saadatmand <nsaadatmand@rigetti.com>", "Tyler Wilson <twilson@rigetti.com>"]
packages = [{ include = "rigetti_resource_estimation", from = "src" }]
include = [
    { path = "CHANGELOG.md", format = ["sdist", "wheel"] },
    { path = "LICENSE", format = ["sdist", "wheel"] },
    { path = "params.yaml", format = ["sdist", "wheel"] },
    { path = "examples", format = ["sdist"] },
]
readme = "README.md"

[tool.poetry.scripts]
sweep = "rigetti_resource_estimation.analysis:sweep"

[tool.poetry.dependencies]
argh = ">=0.26.2"
attrs = ">=24.0"
cirq-core = "=1.4"
graph-state-generation = { git = "https://github.com/sfc-aqua/gosc-graph-state-generation" }
cabaliser = {path = "external/cabaliser", develop = true}
juliacall = "==0.9.12"
mbqc-scheduling = ">=0.1"
networkx = "^2.8.3"
pandas = "^1.5.3"
pauli-tracker = ">=0.1"
python = ">=3.10,<3.12"
qiskit = ">=0.24"
stim = ">=1.10"
pyliqtr = "1.3.3"


[tool.poetry.dev-dependencies]
black = {extras = ["jupyter"], version = "^22.6.0"}
ipywidgets = "^7.7"
jupyterlab = "^3.4"
jupyterlab-lsp = "^3.10"
jupyterlab-git = "^0.34"
papermill = "^2.4"
pytest = "^7.1"
pytest-cov = "^3.0"
pytest-profiling = "^1.7"
pyyaml = "^6.0"
qiskit-aer = ">=0.14.0.1"
ruff = "^0.0.280"
sphinx = "^5.0.0"
sphinx-autodoc-typehints = "^1.17"
sphinx-rtd-theme = "^1.0"

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
line-length = 120
exclude = [
    "external/"
]
